move: LottoPrize 패키지 이동
