+++ src/test/kotlin/lotto/model/LottoTest.kt
@@ -1,6 +1,5 @@
 package lotto.model
 
-import lotto.constants.LottoPrize
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
