+++ src/test/kotlin/lotto/service/ResultCalculatorTest.kt
@@ -1,8 +1,8 @@
 package lotto.service
 
-import lotto.constants.LottoPrize
 import lotto.model.Lotto
 import lotto.model.LottoNumber
+import lotto.model.LottoPrize
 import lotto.model.LottoStore
 import lotto.model.PurchaseInfo
 import lotto.model.WinningLotto
